search and score bug fixes
search: it turns out you can't search for [^WORD], because there are
cases where a guess could include one of those letters, and it would
just be scored 0 (ABACA vs. ALIBI)

score: check if the letter was scored in exact-match before rescoring it
in almost-match, you'll overwrite the exact-match (BAILS vs. ALIBI)