regex each word
It is actually faster to run one regex per remaining word here, because
it saves time on the score comparison afterward. Most words turn up
single-digit numbers of matches per guess pattern. It's faster to
compare those single entries to the word that generated their regex,
than it is to go back through the list looking for that word.