move guess lookup to background webworker
Finding guesses that match patterns can be expensive, like when any of
the match-all patterns are used. This change moves that search to a
background webworker thread, so that the UI will remain responsive. The
answer filtering is still done on the main thread, so the game can
progress while this extra information is acquired.