display the correct "more" (js scoping)
This code attempted to be clever - if the same pattern was entered
multiple times, then when the guesser delivered the list of words for
that pattern, the "preview" element for each pattern would be created in
a loop (as opposed to requiring the guesser to deliver the list multiple
times).

But, the reveal code (preview.onclick) referenced a variable scoped to
the loop. So, no matter which "more" button was clicked, the last list
of words added were the ones that would be revealed.

This fix uses `this.previousElementSibling` instead of the loop-scope
variable. An alternative would have been to capture `rest` in a thunk.