never request guesses for the answer row
The words that fit the answer row are in 'remainingWords'. We don't need
the guess threads time to find them. We also don't want to wait on the
guess thread's time, because the guess count/word hints are not
necessary to figure out when words are ruled in/out by patterns alone.

Also take this opportunity to filter out potentially repeated patterns.
Repeated patterns might be useful for ruing out specific guesses that
generate the same pattern.